feat(test): update sign and normalization tests for new implementation
- Updated EngineMathStandardTest.evalSign to reflect new Number-based sign handling.
- Adjusted normalizeResultTest to validate IntegerResultTest and DoubleResultTest with the updated normalization logic.
- Updated engineMathBinaryExpected to align with the new evalSign and normalizeResult behavior.

+++ app/src/test/java/com/example/calculatorApp/domain/EngineMathStandardTest.kt
@@ -173,7 +173,6 @@ class EngineMathStandardTest {
                 testData.input.previousNumber,
                 testData.input.lastOperator,
             ).value
-
             evalPercent shouldBe testData.expected.value
         }
 
@@ -188,7 +187,6 @@ class EngineMathStandardTest {
                 testData.input.previousNumber,
                 testData.input.lastOperator,
             ).value
-
             evalPercent shouldBe testData.expected.value
         }
 
@@ -203,7 +201,6 @@ class EngineMathStandardTest {
                 testData.input.previousNumber,
                 testData.input.lastOperator,
             ).value
-
             evalPercent shouldBe testData.expected.value
         }
     }
@@ -237,7 +234,7 @@ class EngineMathStandardTest {
             testData: TestCase<InputEngineMath.Sign, ExpectedEngineMath.Sign>
         ) {
             // Act & Assert:
-            val sign = engineMath.evalSign(testData.input.operand.toDouble())
+            val sign = engineMath.evalSign(testData.input.operand)
             sign.value shouldBe testData.expected.value
         }
 
@@ -247,7 +244,7 @@ class EngineMathStandardTest {
             testData: TestCase<InputEngineMath.Sign, ExpectedEngineMath.Sign>
         ) {
             // Act & Assert:
-            val sign = engineMath.evalSign(testData.input.operand.toDouble())
+            val sign = engineMath.evalSign(testData.input.operand)
             sign.value shouldBe testData.expected.value
         }
 
@@ -257,7 +254,7 @@ class EngineMathStandardTest {
             testData: TestCase<InputEngineMath.Sign, ExpectedEngineMath.Sign>
         ) {
             // Act & Assert:
-            val sign = engineMath.evalSign(testData.input.operand.toDouble())
+            val sign = engineMath.evalSign(testData.input.operand)
             sign.value shouldBe testData.expected.value
         }
     }
